refactor: TestAuthClient 설정 추가
