+++ backend/src/main/java/com/zzang/chongdae/auth/service/client/DevAuthClient.java
@@ -1,16 +1,16 @@
-package com.zzang.chongdae.auth.service;
+package com.zzang.chongdae.auth.service.client;
 
 import com.zzang.chongdae.global.exception.MarketException;
 import org.springframework.web.client.RestClient;
 
-public class DevAuthClient extends AuthClient {
+public class DevAuthClient extends ProdAuthClient {
 
     public DevAuthClient(RestClient restClient) {
         super(restClient);
     }
 
     @Override
-    public String getKakaoUserInfo(String accessToken) {
+    protected String getKakaoUserInfo(String accessToken) {
         try {
             return super.getKakaoUserInfo(accessToken);
         } catch (MarketException e) {
