+++ backend/src/test/java/com/zzang/chongdae/auth/service/StubAuthClient.java
@@ -1,0 +1,9 @@
+package com.zzang.chongdae.auth.service;
+
+public class StubAuthClient implements AuthClient {
+
+    @Override
+    public String getUserInfo(String accessToken) {
+        return "stubUserInfo";
+    }
+}
