+++ backend/src/test/java/com/zzang/chongdae/member/service/NicknameGeneratorTest.java
@@ -1,5 +1,6 @@
 package com.zzang.chongdae.member.service;
 
+import com.zzang.chongdae.auth.config.TestAuthClientConfig;
 import com.zzang.chongdae.member.config.TestNicknameWordPickerConfig;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -10,7 +11,8 @@ import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.test.context.ActiveProfiles;
 
 @ActiveProfiles("test")
-@SpringBootTest(webEnvironment = WebEnvironment.NONE, classes = {TestNicknameWordPickerConfig.class})
+@SpringBootTest(webEnvironment = WebEnvironment.NONE, classes = {TestNicknameWordPickerConfig.class,
+        TestAuthClientConfig.class})
 public class NicknameGeneratorTest {
 
     @Autowired
